build: add macOS (Apple Silicon) native build support
Port OvenMediaEngine to build and run natively on macOS arm64.
Key changes include kqueue-based epoll wrapper, POSIX compatibility
fixes (sem_init, pthread_setname_np, BSD sed), clock type alignment
(high_resolution_clock vs system_clock), and Linux-only API guards
(SIOCINQ, TCP_QUICKACK, malloc_trim, etc.).

+++ src/projects/modules/containers/bmff/fmp4_packager/fmp4_structure.h
@@ -136,7 +136,7 @@ namespace bmff
 			return _data;
 		}
 
-		size_t GetDataLength() const override
+		uint64_t GetDataLength() const override
 		{
 			return _data == nullptr ? 0 : _data->GetLength();
 		}
